Move classes to Data namespace
