+++ sources/Google.Solutions.IapDesktop.Extensions.Shell.Test/Services/ConnectionSettings/TestConnectionSettingsRepository.cs
@@ -21,6 +21,7 @@
 
 using Google.Solutions.Common.Locator;
 using Google.Solutions.IapDesktop.Application.Services.Settings;
+using Google.Solutions.IapDesktop.Extensions.Shell.Data;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.Connection;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.ConnectionSettings;
 using Microsoft.Win32;
