+++ sources/Google.Solutions.IapDesktop.Extensions.Shell.Test/Services/ConnectionSettings/TestIapRdpUrlConnectionSettings.cs
@@ -21,6 +21,7 @@
 
 using Google.Solutions.Common.Locator;
 using Google.Solutions.IapDesktop.Application.Data;
+using Google.Solutions.IapDesktop.Extensions.Shell.Data;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.Connection;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.ConnectionSettings;
 using NUnit.Framework;
