+++ sources/Google.Solutions.IapDesktop.Extensions.Shell/Google.Solutions.IapDesktop.Extensions.Shell.csproj
@@ -129,8 +129,8 @@
     <Compile Include="Services\Adapter\KeyStoreAdapter.cs" />
     <Compile Include="Services\Adapter\QuarantineAdapter.cs" />
     <Compile Include="Services\Connection\ConnectionTemplate.cs" />
-    <Compile Include="Services\Connection\RdpSessionParameters.cs" />
-    <Compile Include="Services\Connection\SshSessionParameters.cs" />
+    <Compile Include="Data\RdpSessionParameters.cs" />
+    <Compile Include="Data\SshSessionParameters.cs" />
     <Compile Include="Services\Ssh\IAuthorizedPublicKey.cs" />
     <Compile Include="Services\Ssh\MetadataAuthorizedPublicKey.cs" />
     <Compile Include="Services\Ssh\MetadataAuthorizedPublicKeySet.cs" />
